requirement.txt
transformers
Pillow
torch
zhipuai